# cmd/docs/swagger.json
{"swagger":"2.0","info":{"description":"This is a sample server for a Social Media Application.","title":"SocialWithGO API","termsOfService":"http://swagger.io/terms/","contact":{"name":"API Support","url":"http://www.swagger.io/support","email":"support@swagger.io"},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"}},"basePath":"localhost:8080/v1","paths":{"/authenticate/token":{"post":{"description":"Creates a token for user","consumes":["application/json"],"produces":["application/json"],"tags":["Authentication"],"summary":"Creats a token","parameters":[{"description":"User credentials","name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/main.CreateUserTokenPayload"}}],"responses":{"201":{"description":"token","schema":{"type":"string"}},"400":{"description":"Bad Request","schema":{}},"500":{"description":"Internal Server Error","schema":{}}}}},"/authenticate/user":{"post":{"description":"Registers a new user with the provided details.","consumes":["application/json"],"produces":["application/json"],"tags":["Authentication"],"summary":"Register a new user","parameters":[{"description":"User registration details","name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/main.RegisterUserPayload"}}],"responses":{"201":{"description":"User registered successfully"},"400":{"description":"Bad request","schema":{"$ref":"#/definitions/main.errorResponse"}},"500":{"description":"Internal server error","schema":{"$ref":"#/definitions/main.errorResponse"}}}}},"/feeds":{"get":{"security":[{"BearerAuth":[]}],"description":"Retrieve the feed for the authenticated user with pagination support.","consumes":["application/json"],"produces":["application/json"],"tags":["Feeds"],"summary":"Get User Feed","parameters":[{"maximum":100,"minimum":1,"type":"integer","default":1,"description":"Number of posts to return","name":"limit","in":"query"},{"minimum":0,"type":"integer","default":0,"description":"Number of posts to skip","name":"offset","in":"query"},{"type":"string","default":"desc","description":"Sort order of posts by creation time","name":"sort","in":"query"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/store.PostWithMetadata"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/main.errorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/main.errorResponse"}}}}},"/posts":{"post":{"security":[{"BearerAuth":[]}],"description":"Creates a new post with the provided content, title, and optional tags.","consumes":["application/json"],"produces":["application/json"],"tags":["Posts"],"summary":"Create a new post","parameters":[{"description":"Post Payload","name":"post","in":"body","required":true,"schema":{"$ref":"#/definitions/main.CreatePostPayload"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/store.Post"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/main.errorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/main.errorResponse"}}}}},"/posts/{postID}":{"get":{"security":[{"BearerAuth":[]}],"description":"Retrieves a post by its ID, including associated comments.","produces":["application/json"],"tags":["Posts"],"summary":"Get a post by ID","parameters":[{"type":"integer","description":"Post ID","name":"postID","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/store.Post"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/main.errorResponse"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/main.errorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/main.errorResponse"}}}},"delete":{"security":[{"BearerAuth":[]}],"description":"Deletes a post by its ID.","tags":["Posts"],"summary":"Delete a post by ID","parameters":[{"type":"integer","description":"Post ID","name":"postID","in":"path","required":true}],"responses":{"204":{"description":"No Content"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/main.errorResponse"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/main.errorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/main.errorResponse"}}}},"patch":{"security":[{"BearerAuth":[]}],"description":"Updates a post's title, content, and/or tags by its ID.","consumes":["application/json"],"produces":["application/json"],"tags":["Posts"],"summary":"Update a post by ID","parameters":[{"type":"integer","description":"Post ID","name":"postID","in":"path","required":true},{"description":"Update Payload","name":"post","in":"body","required":true,"schema":{"$ref":"#/definitions/main.UpdatePayload"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/store.Post"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/main.errorResponse"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/main.errorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/main.errorResponse"}}}}},"/users/activate/{token}":{"put":{"description":"Activates a user account using the provided activation token.","tags":["Users"],"summary":"Activate a user account","parameters":[{"type":"string","description":"Activation Token","name":"token","in":"path","required":true}],"responses":{"204":{"description":"No Content"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/main.errorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/main.errorResponse"}}}}},"/users/{userID}":{"get":{"description":"Get a user by their unique ID.","consumes":["application/json"],"produces":["application/json"],"tags":["Users"],"summary":"Get User by ID","parameters":[{"type":"integer","description":"User ID","name":"userID","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/store.Users"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/main.errorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/main.errorResponse"}}}}}},"definitions":{"main.CreatePostPayload":{"type":"object","required":["content","title"],"properties":{"content":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"title":{"type":"string"}}},"main.CreateUserTokenPayload":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","maxLength":255},"password":{"type":"string","maxLength":72,"minLength":3}}},"main.RegisterUserPayload":{"type":"object","required":["email","password","username"],"properties":{"email":{"type":"string"},"password":{"type":"string","minLength":8},"username":{"type":"string","maxLength":30,"minLength":3}}},"main.UpdatePayload":{"type":"object","properties":{"content":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"title":{"type":"string"}}},"main.errorResponse":{"type":"object","properties":{"error":{"type":"string"}}},"store.Comments":{"type":"object","properties":{"content":{"type":"string"},"created_at":{"type":"string"},"id":{"type":"integer"},"post_id":{"type":"integer"},"user":{"$ref":"#/definitions/store.Users"},"user_id":{"type":"integer"}}},"store.Post":{"type":"object","properties":{"comments":{"type":"array","items":{"$ref":"#/definitions/store.Comments"}},"content":{"type":"string"},"created_at":{"type":"string"},"id":{"type":"integer"},"tags":{"type":"array","items":{"type":"string"}},"title":{"type":"string"},"user_id":{"type":"integer"},"username":{"type":"string"},"version":{"type":"integer"}}},"store.PostWithMetadata":{"type":"object","properties":{"comments":{"type":"array","items":{"$ref":"#/definitions/store.Comments"}},"comments_count":{"type":"integer"},"content":{"type":"string"},"created_at":{"type":"string"},"id":{"type":"integer"},"tags":{"type":"array","items":{"type":"string"}},"title":{"type":"string"},"user_id":{"type":"integer"},"username":{"type":"string"},"version":{"type":"integer"}}},"store.Users":{"type":"object","properties":{"created_at":{"type":"string"},"email":{"type":"string"},"id":{"type":"integer"},"is_active":{"type":"boolean"},"username":{"type":"string"}}}},"securityDefinitions":{"BearerAuth":{"description":"Type \"Bearer\" followed by a space and JWT token.","type":"apiKey","name":"Authorization","in":"header"}}}